refactor: move layout store to ui
