+++ apps/main/src/analytics/components/PageHome/components/ComingSoon.tsx
@@ -1,9 +1,9 @@
 import { MIN_HEIGHT } from '@/analytics/features/charts/components/EChartsCard'
-import { EmptyStateCard } from '@evm-ui/shared/ui/EmptyStateCard'
 import Card from '@mui/material/Card'
 import CardContent from '@mui/material/CardContent'
 import CardHeader from '@mui/material/CardHeader'
 import Stack from '@mui/material/Stack'
+import { EmptyStateCard } from '@ui/components/EmptyStateCard'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { CrvHubIcon } from '@ui/icons/CrvHubIcon'
 import { LAFIcon } from '@ui/icons/LAFIcon'
