+++ apps/main/src/bridge/features/bridge/components/BridgeTargets.tsx
@@ -4,7 +4,6 @@ import { ChainList } from '@evm-ui/features/switch-chain/ui/ChainList'
 import { ChainSwitcherIcon } from '@evm-ui/features/switch-chain/ui/ChainSwitcherIcon'
 import { usePathname } from '@evm-ui/hooks/router'
 import { getCurrentNetwork } from '@evm-ui/shared/routes'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import { requireBlockchainId } from '@evm-ui/utils/network'
 import type { NetworkDef } from '@legacy-ui/utils'
 import { ArrowRight } from '@mui/icons-material'
@@ -13,6 +12,7 @@ import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 import { Chain } from '@primitives/network.utils'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { Select } from '@ui/components/Select'
 import { Spinner } from '@ui/components/Spinner'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
