+++ apps/main/src/bridge/features/bridges/components/BridgeOverview.tsx
@@ -1,10 +1,10 @@
 import Fuse from 'fuse.js'
 import { useMemo, useState } from 'react'
 import { LegacyTableSearchField } from '@evm-ui/shared/ui/DataTable/LegacyTableSearchField'
-import { EmptyStateCard } from '@evm-ui/shared/ui/EmptyStateCard'
 import type { Partner } from '@evm-ui/shared/ui/PartnerCard'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
+import { EmptyStateCard } from '@ui/components/EmptyStateCard'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsMobile } from '@ui/hooks/useBreakpoints'
 import { useSwitch } from '@ui/hooks/useSwitch'
