+++ apps/main/src/dao/components/PageGauges/CreateVote/CreateVoteModal.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react'
 import { zeroAddress } from 'viem'
 import { usePinataJwt } from '@evm-ui/hooks/useLocalStorage'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import { FormAlerts } from '@evm-ui/widgets/DetailPageLayout/FormAlerts'
 import Visibility from '@mui/icons-material/Visibility'
 import VisibilityOff from '@mui/icons-material/VisibilityOff'
@@ -15,6 +14,7 @@ import Stack from '@mui/material/Stack'
 import TextField from '@mui/material/TextField'
 import Typography from '@mui/material/Typography'
 import { InlineLink } from '@ui/components/InlineLink'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { FormProvider } from '@ui/features/forms'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
