+++ apps/main/src/dao/components/PageVeCrv/index.tsx
@@ -4,9 +4,9 @@ import { networksIdMapper } from '@/dao/networks'
 import type { NetworkUrlParams } from '@/dao/types/dao.types'
 import { ConnectWalletPrompt } from '@evm-ui/features/connect-wallet'
 import { useParams } from '@evm-ui/hooks/router'
-import { ErrorMessage } from '@evm-ui/shared/ui/ErrorMessage'
 import { FormSkeleton } from '@evm-ui/widgets/DetailPageLayout/FormSkeleton'
 import Box from '@mui/material/Box'
+import { ErrorMessage } from '@ui/features/errors/ErrorMessage'
 import { FormPlacementProvider } from '@ui/features/form-context/FormPlacementProvider'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
