+++ apps/main/src/lend/components/PageLendMarket/LendMarketPage.tsx
@@ -26,11 +26,11 @@ import {
   useMarketMobileFormDrawer,
   useNewLlamaMarketDetailPage,
 } from '@evm-ui/hooks/useFeatureFlags'
-import { ErrorPage } from '@evm-ui/pages/ErrorPage'
 import { MarketType, MarketRateType } from '@evm-ui/types/market'
 import { DetailPageLayout } from '@evm-ui/widgets/DetailPageLayout/DetailPageLayout'
 import { DetailPageSection as MarketSection } from '@evm-ui/widgets/DetailPageLayout/DetailPageSection'
 import type { Decimal } from '@primitives/decimal.utils'
+import { ErrorPage } from '@ui/features/errors/ErrorPage'
 import type { Range } from '@ui/features/queries/util'
 import { t } from '@ui/lib/i18n'
 import { useLendMarket } from '../../hooks/useLendMarket'
