+++ apps/main/src/llamalend/features/market-list/UserPositionsTables.tsx
@@ -4,12 +4,12 @@ import type { LlamaMarketsTableResult } from '@/llamalend/queries/market-list/ll
 import type { LlamaMarket } from '@/llamalend/queries/market-list/llama-markets'
 import { getInternalUrl, LEND_MARKET_ROUTES, LEND_ROUTES } from '@evm-ui/shared/routes'
 import { TableHeader } from '@evm-ui/shared/ui/DataTable/TableHeader'
-import { EmptyStateCard } from '@evm-ui/shared/ui/EmptyStateCard'
-import { ErrorMessage } from '@evm-ui/shared/ui/ErrorMessage'
 import { MarketRateType } from '@evm-ui/types/market'
 import Stack from '@mui/material/Stack'
 import type { Address } from '@primitives/address.utils'
 import { fromEntries, maybe, recordValues } from '@primitives/objects.utils'
+import { EmptyStateCard } from '@ui/components/EmptyStateCard'
+import { ErrorMessage } from '@ui/features/errors/ErrorMessage'
 import { mapQuery, QueryProp } from '@ui/features/queries/util'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
