+++ apps/main/src/llamalend/widgets/action-card/LowSolvencyActionModal.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import { CheckboxField } from '@evm-ui/widgets/DetailPageLayout/CheckboxField'
 import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useSwitch } from '@ui/hooks/useSwitch'
 import { t } from '@ui/lib/i18n'
