+++ apps/main/src/loan/components/PageMintMarket/MintMarketPage.tsx
@@ -21,11 +21,11 @@ import { useCurve } from '@evm-ui/features/connect-wallet'
 import { useUserProfileStore } from '@evm-ui/features/user-profile'
 import { useParams } from '@evm-ui/hooks/router'
 import { useMarketMobileFormDrawer, useNewLlamaMarketDetailPage } from '@evm-ui/hooks/useFeatureFlags'
-import { ErrorPage } from '@evm-ui/pages/ErrorPage'
 import { MarketType, MarketRateType } from '@evm-ui/types/market'
 import { DetailPageLayout } from '@evm-ui/widgets/DetailPageLayout/DetailPageLayout'
 import { DetailPageSection as MarketSection } from '@evm-ui/widgets/DetailPageLayout/DetailPageSection'
 import type { Decimal } from '@primitives/decimal.utils'
+import { ErrorPage } from '@ui/features/errors/ErrorPage'
 import type { Range } from '@ui/features/queries/util'
 import { t } from '@ui/lib/i18n'
 import { useMintMarket } from '../../hooks/useMintMarket'
