+++ apps/main/src/main.tsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import { ErrorBoundary } from '@evm-ui/widgets/ErrorBoundary'
 import MuiLink from '@mui/material/Link'
 import { RouterProvider } from '@tanstack/react-router'
+import { ErrorBoundary } from '@ui/features/errors/ErrorBoundary'
 import { initSentry } from '@ui/features/sentry'
 import { t } from '@ui/lib/i18n'
 import { router } from './routes'
