+++ apps/main/src/routes/RootLayout.tsx
@@ -18,13 +18,13 @@ import { useLayoutStoreResponsive } from '@evm-ui/hooks/useLayoutStoreResponsive
 import { useNetworkFromUrl } from '@evm-ui/hooks/useNetworkFromUrl'
 import { useOnChainUnavailable } from '@evm-ui/hooks/useOnChainUnavailable'
 import { getCurrentApp } from '@evm-ui/shared/routes'
-import { ErrorBoundary } from '@evm-ui/widgets/ErrorBoundary'
 import MuiLink from '@mui/material/Link'
 import { maybe, recordValues } from '@primitives/objects.utils'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { HeadContent, Outlet } from '@tanstack/react-router'
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
 import { ThemeProvider } from '@ui/components/ThemeProvider'
+import { ErrorBoundary } from '@ui/features/errors/ErrorBoundary'
 import { QueryProvider } from '@ui/features/queries/provider'
 import { persister, queryClient } from '@ui/features/queries/query-client'
 import { addBreadcrumb } from '@ui/features/sentry'
