+++ packages/evm-ui/src/features/connect-wallet/ui/WagmiConnectModal.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useMemo, useState } from 'react'
 import type { BaseError } from 'viem'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import Alert from '@mui/material/Alert'
 import AlertTitle from '@mui/material/AlertTitle'
 import Box from '@mui/material/Box'
@@ -8,6 +7,7 @@ import MenuList from '@mui/material/MenuList'
 import { createSvgIcon } from '@mui/material/utils'
 import { toArray } from '@primitives/array.utils'
 import { MenuItem } from '@ui/components/MenuItem'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { handleBreakpoints } from '@ui/features/themes/basic-theme'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { BrowserWalletIcon } from '@ui/icons/BrowserWalletIcon'
