+++ packages/evm-ui/src/features/maintenance/components/BackendMaintenanceModal.tsx
@@ -1,9 +1,9 @@
 import { ReactNode } from 'react'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import MuiTypography from '@mui/material/Typography'
 import { ExternalLink } from '@ui/components/ExternalLink'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
 import { Maintenance } from '../hooks/useMaintenance'
