+++ packages/evm-ui/src/features/submit-error-report/ErrorReportModal.stories.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react'
 import { http, zeroAddress } from 'viem'
 import { mainnet } from 'viem/chains'
 import { WagmiProvider } from 'wagmi'
-import { ErrorReportModal } from '@evm-ui/features/report-error'
 import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import type { Meta, StoryObj } from '@storybook/react-vite'
+import { ErrorReportModal } from '@ui/features/report-error'
 import { t } from '@ui/lib/i18n'
 import { createConfig } from '@wagmi/core'
