+++ packages/evm-ui/src/features/user-profile/settings/ReleaseChannelDialog.tsx
@@ -1,9 +1,9 @@
 import { showReleaseChannelSnackbar } from '@evm-ui/features/user-profile/settings/settings.util'
 import { useReleaseChannel } from '@evm-ui/hooks/useLocalStorage'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t, Trans } from '@ui/lib/i18n'
 import { ReleaseChannel } from '@ui/utils/env'
