+++ packages/evm-ui/src/shared/ui/Chart/ChartEmpty.tsx
@@ -1,8 +1,8 @@
 import type { ReactNode } from 'react'
 import Stack from '@mui/material/Stack'
+import { EmptyStateCard } from '@ui/components/EmptyStateCard'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
-import { EmptyStateCard } from '../EmptyStateCard'
 
 const { Spacing } = SizesAndSpaces
