+++ packages/evm-ui/src/shared/ui/DataTable/DataTable.tsx
@@ -11,12 +11,12 @@ import TableFooter from '@mui/material/TableFooter'
 import TableHead from '@mui/material/TableHead'
 import TableRow from '@mui/material/TableRow'
 import type { RowData } from '@tanstack/react-table'
+import { EmptyStateCard, type EmptyStateCardProps } from '@ui/components/EmptyStateCard'
 import { WithWrapper } from '@ui/components/WithWrapper'
 import { useLayoutStore } from '@ui/features/layout/layout'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsMobile } from '@ui/hooks/useBreakpoints'
 import { t } from '@ui/lib/i18n'
-import { EmptyStateCard, EmptyStateCardProps } from '../EmptyStateCard'
 import { DATA_TABLE_CATEGORIES, type DataTableCategory, type DataTableCategoryConfig } from './categories'
 import { DataTableHeaderHeight, type useCurveTable } from './data-table.utils'
 import { DataRow, type DataRowProps } from './DataRow'
