+++ packages/evm-ui/src/shared/ui/EvmErrorMessage.tsx
@@ -1,6 +1,6 @@
 import type { ComponentProps } from 'react'
 import { useConnection } from 'wagmi'
-import { ErrorMessage } from '@evm-ui/shared/ui/ErrorMessage'
+import { ErrorMessage } from '@ui/features/errors/ErrorMessage'
 
 export const EvmErrorMessage = (props: Omit<ComponentProps<typeof ErrorMessage>, 'userAddress'>) => (
   <ErrorMessage {...props} userAddress={useConnection().address} />
