+++ packages/evm-ui/src/utils/index.ts
@@ -6,7 +6,7 @@ export * from './network'
 export * from './number'
 export * from './pagination'
 export * from './decimal'
-export * from './errors'
+export * from '@ui/features/errors/errors.util'
 export * from './average-categories'
 export * from './rates'
 export * from './tokens'
