+++ packages/evm-ui/src/widgets/DetailPageLayout/FormAlerts.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react'
-import { ErrorReportModal } from '@evm-ui/features/report-error'
 import { usePreviousValue } from '@evm-ui/hooks/usePreviousValue'
 import { CopyIconButton } from '@evm-ui/shared/ui/CopyIconButton'
 import { formatNumber, getErrorMessage } from '@evm-ui/utils'
@@ -21,6 +20,7 @@ import type { Decimal } from '@primitives/decimal.utils'
 import { maybe } from '@primitives/objects.utils'
 import { WithSkeleton } from '@ui/components/WithSkeleton'
 import { type QueryProp } from '@ui/features/queries/util'
+import { ErrorReportModal } from '@ui/features/report-error'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useSwitch } from '@ui/hooks/useSwitch'
 import { t } from '@ui/lib/i18n'
