+++ packages/evm-ui/src/widgets/EvmErrorBoundary.tsx
@@ -1,6 +1,6 @@
 import type { ComponentProps } from 'react'
 import { useConnection } from 'wagmi'
-import { ErrorBoundary } from '@evm-ui/widgets/ErrorBoundary'
+import { ErrorBoundary } from '@ui/features/errors/ErrorBoundary'
 
 export const EvmErrorBoundary = (props: Omit<ComponentProps<typeof ErrorBoundary>, 'userAddress'>) => (
   <ErrorBoundary {...props} userAddress={useConnection().address} />
