+++ packages/evm-ui/src/widgets/SlippageSettings/SlippageSettingsModal.tsx
@@ -1,10 +1,10 @@
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import { FormAlerts } from '@evm-ui/widgets/DetailPageLayout/FormAlerts'
 import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import type { Address } from '@primitives/address.utils'
 import { toArray } from '@primitives/array.utils'
 import type { Decimal } from '@primitives/decimal.utils'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { FormProvider } from '@ui/features/forms'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
