+++ packages/ui/src/features/report-error/ErrorReportModal.tsx
@@ -1,5 +1,4 @@
 import { type MouseEvent, useEffect } from 'react'
-import { ModalDialog } from '@evm-ui/shared/ui/ModalDialog'
 import Button from '@mui/material/Button'
 import FormControl from '@mui/material/FormControl'
 import FormHelperText from '@mui/material/FormHelperText'
@@ -10,6 +9,7 @@ import ToggleButton from '@mui/material/ToggleButton'
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup'
 import Typography from '@mui/material/Typography'
 import type { Address } from '@primitives/address.utils'
+import { ModalDialog } from '@ui/components/ModalDialog'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { t } from '@ui/lib/i18n'
 import { type ContactMethod, type ErrorContext, useErrorReportForm } from './useErrorReportForm'
