+++ packages/ui/src/lib/timers.spec.ts
@@ -1,6 +1,6 @@
 import { noop } from 'lodash'
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { setTimeoutInterval } from './timers'
+import { setTimeoutInterval } from '@ui/lib/timers'
 
 const createDeferred = <T = void>() => {
   let resolve!: (value: T | PromiseLike<T>) => void
